refactor(사용자 입력 기능 - 메뉴 목록 입력 기능) : 커스텀 예외 패키지 변경
